requirements.txt
# MCP Server Framework
mcp>=1.0.0

# Text processing and AI
openai>=1.0.0
anthropic>=0.25.0
requests>=2.31.0
urllib3>=1.26.0
prometheus_client>=0.19.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
jinja2>=3.1.0
markdown>=3.5.0
pyyaml>=6.0.0

# Date/time handling
python-dateutil>=2.8.0